refactor(auth): relocate get_current_user to app/lib/auth.py
