+++ app/lib/auth.py
@@ -1,0 +1,51 @@
+"""Authentication dependency for resolving the current user from a bearer token.
+
+Public API home for ``get_current_user`` (plugins and application code import it here).
+``app.api.v1.auth`` re-exports it for backward compatibility.
+"""
+
+import jwt
+from fastapi import Depends, HTTPException, status
+from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from sqlmodel import select
+from sqlmodel.ext.asyncio.session import AsyncSession
+
+from app.core import security
+from app.lib.db import get_async_session
+from app.models.user import User
+
+security_scheme = HTTPBearer()
+
+
+async def get_current_user(
+    credentials: HTTPAuthorizationCredentials = Depends(security_scheme),
+    session: AsyncSession = Depends(get_async_session),
+) -> User:
+    token = credentials.credentials
+
+    try:
+        payload = security.decode_access_token(token)
+        username = payload.get("sub")
+        if not isinstance(username, str) or username is None:
+            raise HTTPException(
+                status_code=status.HTTP_401_UNAUTHORIZED,
+                detail="Could not validate credentials",
+                headers={"WWW-Authenticate": "Bearer"},
+            )
+    except jwt.InvalidTokenError:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
+    result = await session.exec(select(User).where(User.username == username))
+    user = result.one_or_none()
+    if user is None:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="User not found",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+
+    return user
